requirements.txt
wheel
FastAPI[all]
gunicorn
uvicorn
Jinja2
minify-html
python-decouple